Fix multiple assignments with self setters.
There is no way to add trackers to a potentially private setter.
We return a conservative `flow_entry_count`, such that these setters are
considered fully executed only if they are, but not "if and only if".
False negatives may be returned, and only if another nearby setter was not fully executed.